[PARQUET-1801] Add parquet-tools 'prune' to parquet-cli
